:truck: restructure the `core` package
