+++ core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod disprove;
 pub mod assert;
 pub mod debug;
 pub mod utils;
